sponge fully synthesizable - with no optimization

+++ hardware/action_examples/hls_sponge/keccak.h
@@ -7,6 +7,7 @@
 #include <stdint.h>
 #include <string.h>
 
+
 #ifndef KECCAK_ROUNDS
 #define KECCAK_ROUNDS 24
 #endif
@@ -16,7 +17,7 @@
 #endif
 
 // compute a keccak hash (md) of given byte length from "in"
-int keccak(const uint8_t *in, int inlen, uint8_t *md, int mdlen);
+int keccak(const uint64_t *in, int inlen, uint64_t *md, int mdlen);
 
 // update the state
 void keccakf(uint64_t st_in[25], uint64_t st_out[25], int norounds);
